project-first/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "enterprise-rag"
version = "0.1.0"
description = "Enterprise RAG System with AI Agent Governance"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "tenacity>=8.0",
    "qdrant-client>=1.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "behave>=1.2",
]
mcp = [
    "mcp>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (Qdrant, OpenAI)",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
include = ["src*"]
